# backend/src/api/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Poker-API
  description: Progrendszerek projekt backend apija (Poker oldal no money)
...
paths:
...
  /user/{userId}:
    get:
      summary: Get user information by id
      operationId: getUserById
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The id of the user to retrieve
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized, user not logged in
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
  /users:
    get:
      summary: Get user information
      operationId: getUsers
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          required: false
          description: The id of the user to retrieve
        - in: query
          name: username
          schema:
            type: string
          required: false
          description: The username of the user to retrieve
          # TODO: paginate?
      responses:
        '200':
          description: User information retrieved successfully
          # TODO: content
        '401':
          description: Unauthorized, user not logged in
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
    put:
      summary: Update user information
      operationId: updateUser
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          required: true
          description: The id of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserChangeRequest'
      responses:
        '200':
          description: User information updated successfully
        '400':
          description: Invalid request
        '401':
          description: Unauthorized, user not logged in
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - userSpecificCookieAuth: []
        - cookieAdminAuth: []
    delete:
      summary: Delete user
      operationId: deleteUser
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          required: true
          description: The id of the user
      responses:
        '200':
          description: User deleted successfully
        '401':
          description: Unauthorized, user not logged in
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - userSpecificCookieAuth: []
        - cookieAdminAuth: []
  /friends:
    get:
      summary: Get friends of user
      operationId: getFriends
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      responses:
        '200':
          description: Friends retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    username:
                      type: string
                    roles:
                      type: array
                      items:
                        type: string
                    balance:
                      type: number
                    createdAt:
                      type: string
                      format: date-time
                    updatedAt:
                      type: string
                      format: date-time
        '401':
          description: Unauthorized, user not logged in
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
    post:
      summary: Add friend to user
      operationId: addFriend
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                friendId:
                  type: string
              required:
                - friendId
      responses:
        '200':
          description: Friend added successfully
        '401':
          description: Unauthorized, user not logged in
        '404':
          description: User or friend not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
    delete:
      summary: Remove friend from user
      operationId: removeFriend
      x-eov-operation-handler: api/controllers/user.controller
      tags:
        - Users
      parameters:
        - in: query
          name: friendId
          schema:
            type: string
          required: true
          description: The id of the user to remove as friend
      responses:
        '200':
          description: Friend removed successfully
        '401':
          description: Unauthorized, user not logged in
        '404':
          description: User or friend not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
# Game
...
components:
  schemas:
...
    LoginResponse:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        balance:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
        - _id
    UserChangeRequest:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        hashedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        balance:
          type: number
      additionalProperties: false
      required:
        - _id
...
    BaseError:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        status:
          type: number
        originatedFrom:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    userSpecificCookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    chatOwnerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    chatMemberAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    gameOwnerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    gamePlayerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    cookieAdminAuth:
      type: apiKey
      in: cookie
      name: connect.sid

# TODO: mindehol jo response
# TODO: error messagek ahol lehet
# TODO: example es description
# TODO: plural or not plural, also generic all endpoint or single endpoint?
# FIXME: change query to path param where needed